requirements.txt
streamlit
pandas
pytesseract
pdfplumber
Pillow
plotly
numpy

